refactor: extract scatterplot data transformation logic
Move data processing into src/lib/scatterplot-data.ts.
