+++ src/components/vulnerability-scatterplot.tsx
@@ -22,15 +22,11 @@ import {
   ToggleGroupItem,
 } from "@/components/ui/toggle-group"
 import { useScans } from "@/hooks/useScans"
-
-interface ScatterplotDataPoint {
-  library: string
-  severity: 'critical' | 'high' | 'medium' | 'low'
-  count: number
-  x: number
-  y: number
-  color: string
-}
+import {
+  transformScansToChartData,
+  generateChartKey,
+  type ScatterplotDataPoint,
+} from "@/lib/scatterplot-data"
 
 export function VulnerabilityScatterplot() {
   const router = useRouter()
@@ -73,163 +69,9 @@ export function VulnerabilityScatterplot() {
     fetchAllClassifications();
   }, [scans]);
 
-  // Helper function to adjust vulnerability counts based on false positive classifications
-  const adjustVulnerabilityCounts = React.useCallback((imageName: string, severities: any) => {
-    const classifications = consolidatedClassifications.get(imageName) || [];
-    const falsePositiveCount = classifications.filter(c => c.isFalsePositive).length;
-    
-    // Log all vulnerabilities for this image
-    console.group(`🔍 Image: ${imageName}`);
-    console.log(`📋 Raw vulnerabilities:`, severities);
-    console.log(`🎯 Total classifications:`, classifications.length);
-    console.log(`❌ False positives:`, falsePositiveCount);
-    
-    if (classifications.length > 0) {
-      console.log(`📊 All classifications for ${imageName}:`, classifications.map(c => ({
-        cveId: c.cveId,
-        severity: c.severity,
-        isFalsePositive: c.isFalsePositive,
-        justification: c.justification
-      })));
-      
-      const falsePositives = classifications.filter(c => c.isFalsePositive);
-      if (falsePositives.length > 0) {
-        console.log(`🚫 False positive CVEs:`, falsePositives.map(fp => ({
-          cveId: fp.cveId,
-          severity: fp.severity,
-          justification: fp.justification
-        })));
-      }
-    }
-    
-    if (falsePositiveCount === 0) {
-      console.log(`✅ No false positives - using original counts`);
-      console.groupEnd();
-      return severities;
-    }
-
-    // Apply a conservative reduction based on false positive classifications
-    // Since we know nginx has 4 critical CVEs all marked as false positive, we'll handle this specifically
-    const totalVulns = (severities.crit || 0) + (severities.high || 0) + (severities.med || 0) + (severities.low || 0);
-    
-    if (totalVulns === 0) {
-      console.groupEnd();
-      return severities;
-    }
-    
-    // More aggressive reduction for critical vulnerabilities since they're most likely to be false positives
-    const criticalReduction = Math.min(falsePositiveCount, severities.crit || 0);
-    const remainingFalsePositives = Math.max(0, falsePositiveCount - criticalReduction);
-    
-    const highReduction = Math.min(remainingFalsePositives, severities.high || 0);
-    const stillRemainingFalsePositives = Math.max(0, remainingFalsePositives - highReduction);
-    
-    const mediumReduction = Math.min(stillRemainingFalsePositives * 0.5, severities.med || 0);
-    const lowReduction = Math.min(stillRemainingFalsePositives * 0.2, severities.low || 0);
-    
-    const adjusted = {
-      crit: Math.max(0, (severities.crit || 0) - criticalReduction),
-      high: Math.max(0, (severities.high || 0) - highReduction),
-      med: Math.max(0, Math.round((severities.med || 0) - mediumReduction)),
-      low: Math.max(0, Math.round((severities.low || 0) - lowReduction)),
-    };
-
-    // Log the adjustment for debugging
-    console.log(`📊 Adjusted vulnerabilities:`, adjusted);
-    console.log(`🔄 Reductions - Critical: ${criticalReduction}, High: ${highReduction}, Medium: ${mediumReduction.toFixed(1)}, Low: ${lowReduction.toFixed(1)}`);
-    console.groupEnd();
-    
-    return adjusted;
-  }, [consolidatedClassifications]);
-
   const chartData = React.useMemo(() => {
-    if (!scans || scans.length === 0) return []
-
-
-    // Group by image name to avoid duplicates from multiple scans of same image
-    // Use the latest scan data for each image (scans are sorted by date DESC)
-    const imageMap = new Map<string, any>()
-    
-    scans.forEach((scan: any) => {
-      if (!scan.severities) return
-      
-      const imageName = scan.imageName || scan.image || 'unknown'
-      
-      // Only keep the first scan for each image (which is the latest due to sorting)
-      if (!imageMap.has(imageName)) {
-        imageMap.set(imageName, scan)
-      }
-    })
-
-    // Create data points from unique images
-    const dataPoints: ScatterplotDataPoint[] = []
-    
-    imageMap.forEach((scan, imageName) => {
-      // Apply false positive adjustments to the vulnerability counts
-      const adjustedSeverities = adjustVulnerabilityCounts(imageName, scan.severities)
-      const { crit, high, med, low } = adjustedSeverities
-      
-      // Create a deterministic position based on image name to ensure consistency
-      const generatePosition = (imageName: string, severityOffset: number) => {
-        let hash = 0;
-        for (let i = 0; i < imageName.length; i++) {
-          const char = imageName.charCodeAt(i);
-          hash = ((hash << 5) - hash) + char;
-          hash = hash & hash; // Convert to 32-bit integer
-        }
-        const normalizedHash = Math.abs(hash) / 2147483647; // Normalize to 0-1
-        const jitter = normalizedHash * 0.15; // Reduced jitter for better spacing
-        return jitter + severityOffset;
-      }
-      
-      // Create data points for each severity level with counts
-      if (crit > 0) {
-        dataPoints.push({
-          library: imageName,
-          severity: 'critical',
-          count: crit,
-          x: generatePosition(imageName, 0.8), // Critical range: 0.8-0.95
-          y: crit,
-          color: '#ef4444'
-        })
-      }
-      
-      if (high > 0) {
-        dataPoints.push({
-          library: imageName,
-          severity: 'high',
-          count: high,
-          x: generatePosition(imageName, 0.6), // High range: 0.6-0.75
-          y: high,
-          color: '#f97316'
-        })
-      }
-      
-      if (med > 0) {
-        dataPoints.push({
-          library: imageName,
-          severity: 'medium',
-          count: med,
-          x: generatePosition(imageName, 0.4), // Medium range: 0.4-0.55
-          y: med,
-          color: '#eab308'
-        })
-      }
-      
-      if (low > 0) {
-        dataPoints.push({
-          library: imageName,
-          severity: 'low',
-          count: low,
-          x: generatePosition(imageName, 0.2), // Low range: 0.2-0.35
-          y: low,
-          color: '#3b82f6'
-        })
-      }
-    })
-    
-    return dataPoints.sort((a, b) => b.count - a.count)
-  }, [scans, adjustVulnerabilityCounts])
+    return transformScansToChartData(scans || [], consolidatedClassifications)
+  }, [scans, consolidatedClassifications])
 
   // Filter chart data based on visible severities
   const filteredChartData = React.useMemo(() => {
@@ -238,12 +80,7 @@ export function VulnerabilityScatterplot() {
 
   // Generate a stable key based on filtered chart data to force re-render when data changes
   const chartKey = React.useMemo(() => {
-    if (filteredChartData.length === 0) return 'empty'
-    const signature = filteredChartData
-      .map(item => `${item.library}:${item.severity}:${item.count}`)
-      .sort()
-      .join('|')
-    return `chart-${filteredChartData.length}-${signature.slice(0, 50)}`
+    return generateChartKey(filteredChartData)
   }, [filteredChartData])
 
   const handlePointClick = React.useCallback((data: ScatterplotDataPoint) => {
@@ -271,7 +108,7 @@ export function VulnerabilityScatterplot() {
   const CustomDot = React.useCallback((props: any) => {
     const { cx, cy, payload } = props
     if (!payload) return <circle cx={cx} cy={cy} r={0} />
-    
+
     const radius = Math.min(Math.max(payload.count * 3 + 4, 8), 24)
     const color = payload.color
     // Create lighter version of the border color
@@ -287,7 +124,7 @@ export function VulnerabilityScatterplot() {
         (B < 255 ? B < 1 ? 0 : B : 255)).toString(16).slice(1)
     }
     const fillColor = lightenColor(color)
-    
+
     return (
       <circle
         cx={cx}
@@ -334,9 +171,9 @@ export function VulnerabilityScatterplot() {
         {/* Findings Toggle Group */}
         <div className="mb-4 flex items-center justify-between">
           <div className="flex-grow"/>
-          <ToggleGroup 
-            type="multiple" 
-            value={visibleSeverities} 
+          <ToggleGroup
+            type="multiple"
+            value={visibleSeverities}
             onValueChange={setVisibleSeverities}
             variant="outline"
             className="justify-end"
@@ -345,8 +182,8 @@ export function VulnerabilityScatterplot() {
               <ToggleGroupItem
                 value="critical"
                 className={`${
-                  visibleSeverities.includes('critical') 
-                    ? '!bg-red-500 !text-white ' 
+                  visibleSeverities.includes('critical')
+                    ? '!bg-red-500 !text-white '
                     : '!bg-white text-red-700 dark:text-red-400 border-red-200 dark:border-red-800'
                 } hover:bg-red-500/20 px-2 py-1 text-xs cursor-pointer`}
               >
@@ -357,8 +194,8 @@ export function VulnerabilityScatterplot() {
               <ToggleGroupItem
                 value="high"
                 className={`${
-                  visibleSeverities.includes('high') 
-                    ? '!bg-orange-500 !text-white' 
+                  visibleSeverities.includes('high')
+                    ? '!bg-orange-500 !text-white'
                     : '!bg-white text-orange-700 dark:text-orange-400 border-orange-200 dark:border-orange-800'
                 } hover:bg-orange-500/20 px-2 py-1 text-xs cursor-pointer`}
               >
@@ -369,8 +206,8 @@ export function VulnerabilityScatterplot() {
               <ToggleGroupItem
                 value="medium"
                 className={`${
-                  visibleSeverities.includes('medium') 
-                    ? '!bg-yellow-500 !text-white ' 
+                  visibleSeverities.includes('medium')
+                    ? '!bg-yellow-500 !text-white '
                     : '!bg-white text-yellow-700 dark:text-yellow-400 border-yellow-200 dark:border-yellow-800'
                 } hover:bg-yellow-500/20 px-2 py-1 text-xs cursor-pointer`}
               >
@@ -381,8 +218,8 @@ export function VulnerabilityScatterplot() {
               <ToggleGroupItem
                 value="low"
                 className={`${
-                  visibleSeverities.includes('low') 
-                    ? '!bg-blue-500 !text-white ' 
+                  visibleSeverities.includes('low')
+                    ? '!bg-blue-500 !text-white '
                     : '!bg-white text-blue-700 dark:text-blue-400 border-blue-200 dark:border-blue-800'
                 } hover:bg-blue-500/20 px-2 py-1 text-xs cursor-pointer`}
               >
@@ -398,7 +235,7 @@ export function VulnerabilityScatterplot() {
               color: "#dc2626",
             },
             high: {
-              label: "High", 
+              label: "High",
               color: "#ea580c",
             },
           }}
@@ -450,7 +287,7 @@ export function VulnerabilityScatterplot() {
         </ChartContainer>
         {filteredChartData.length === 0 && (
           <div className="text-center text-muted-foreground text-sm mt-4">
-            {chartData.length === 0 
+            {chartData.length === 0
               ? "No vulnerabilities found across scanned images"
               : "No vulnerabilities found for the selected severity levels"
             }
